Message: Modified rename file src/pages/admin/dashboard/Index. src/pages/admin/detail/Index. src/pages/admin/products/Index.tsx
