+++ src/pages/admin/detail/Index.tsx
@@ -1,6 +1,6 @@
 export default function Detail() {
-  const 
-    return (
-
+  return (
+    <>
+    </>
   )
 }
