+++ src/router.tsx
@@ -1,13 +1,12 @@
 import { createBrowserRouter } from "react-router-dom";
 import MainLayout from "./components/layouts/MainLayout";
 import Home from "./pages/guest/Home";
-import AdminProducts from "./pages/admin/products/Products"
+import AdminProducts from "./pages/admin/products/Index"
 import DashboardLayout from "./components/layouts/DashboardLayout";
-import Dashboard from "./pages/admin/dashboard/Dashboard";
+import Dashboard from "./pages/admin/dashboard/Index";
 import AuthLayouts from "./components/layouts/AuthLayouts";
 import Login from "./pages/auth/login/Index";
 import Register from "./pages/auth/register/Index";
-import Etalase from "./pages/guest/Etalase";
 
 export const router = createBrowserRouter([
     {
